Stop copying data that already exists in program storage

+++ external_modules/registry_schemas/rgb_led.c
@@ -7,7 +7,7 @@
 #include "registry.h"
 #include "registry_schemas.h"
 
-static void mapping(int param_id, registry_instance_t *instance, void **val, int *val_len);
+static void mapping(int param_id, registry_instance_t *instance, void **val, size_t *val_len);
 
 REGISTRY_SCHEMA(
     registry_schema_rgb_led,
@@ -29,7 +29,7 @@ REGISTRY_SCHEMA(
 
     );
 
-static void mapping(int param_id, registry_instance_t *instance, void **val, int *val_len)
+static void mapping(int param_id, registry_instance_t *instance, void **val, size_t *val_len)
 {
     registry_schema_rgb_led_t *_instance = (registry_schema_rgb_led_t *)instance->data;
